pyproject.toml
[project]
name = "deploydocus2"
version = "0.1.0"
description = ""
authors = [
    {name = "Partho Bhowmick",email = "p.s.bhowmick@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "kubernetes-asyncio-pydantic @ file:///Users/partho/kubernetes-python-client-project-root/python-async-client/dist/kubernetes_asyncio_pydantic-29.1.0-py3-none-any.whl",
]
license = "AGPL-3.0-or-later"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Typing :: Typed",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
mypy = "^1.15.0"
flake8 = "^7.1.2"
black = "^25.1.0"
wheel = "^0.45.1"
cython = "^3.0.12"
flake8-pydantic = "^0.4.0"
setuptools = "^75.8.2"
pip = "^25.0.1"


[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.7"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[tool.black]
line-length = 88
target-version = ["py312", "py313"]

[tool.isort]
profile = "black"
